requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Auth & Security ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
cryptography==42.0.8
bcrypt==4.0.1

# ── Config ───────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# ── HTTP Client (integration tests) ─────────────────────────────────────────
httpx==0.27.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
behave==1.2.6

# ── Linting ──────────────────────────────────────────────────────────────────
ruff==0.4.4
